refactor(e2e): extract coverage engine into e2e-coverage/
Split the two ~1,000-line files (scripts/e2e-coverage.mjs,
scenario-matrix.mjs) into utils/, parsers/, catalog/, scoring/, and
report/ modules. CLI flags and output are unchanged. Old import paths
stay as thin shims.
