+++ frontend/test-e2e/e2e-coverage/catalog/id-prefixes.mjs
@@ -1,0 +1,29 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+/** First part of a case id: which area of the app it covers. */
+
+export const ID_PREFIXES = {
+  L: "Landing",
+  A: "Authentication",
+  H: "Home",
+  E: "Events catalog",
+  EA: "Event about",
+  EF: "Event FAQ",
+  ER: "Event resources",
+  EC: "Event create (modal)",
+  O: "Organizations catalog",
+  OA: "Organization about",
+  OF: "Organization FAQ",
+  OR: "Organization resources",
+  OE: "Organization events",
+  OG: "Organization groups list",
+  OC: "Organization create (modal)",
+  GA: "Group about",
+  GF: "Group FAQ",
+  GR: "Group resources",
+  GE: "Group events",
+  GC: "Group create (modal)",
+  S: "Global search",
+  GG: "Global groups catalog",
+  C: "Contact",
+  T: "Route query / topics",
+};
