+++ frontend/test-e2e/e2e-coverage/scoring/constants.mjs
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+/**
+ * Required-scenario coverage the suite must keep. Raise this as gaps close;
+ * enforced by test-e2e/specs/all/e2e-coverage/e2e-scenario-coverage.spec.ts.
+ */
+export const MIN_REQUIRED_SCENARIO_PCT = 85;
